feat: 볼,스트라이크 관련 테스트케이스 추가
